refactor raw list scroll hook
